requirements.txt
django==3.1.5
djangorestframework==3.12.2
djangorestframework_simplejwt==4.6.0
django-rest-framework-social-oauth2==1.1.0
celery==4.4.7
redis==3.5.3
django-celery-results==1.2.1
celery-progress==0.0.10
channels==3.0.3
channels-redis==3.2.0
pytest-asyncio==0.14.0
django-heroku==0.3.1
whitenoise==5.2.0
git+ssh://git@github.com/GustavoRonconi/simplecapp_calculation_engine.git@0.1.4
